+++ server/src-rsr/migrations/43_to_44.sql
@@ -1,4 +1,5 @@
 -- This migration adds versioning to metadata, used for optimistic locking in UIs.
 -- TODO: Are there changes required in catalog_versions.txt
 
-ALTER TABLE hdb_catalog.hdb_metadata ADD COLUMN "resource_version" INTEGER NOT NULL DEFAULT 1 UNIQUE;
+ALTER TABLE hdb_catalog.hdb_metadata
+ADD COLUMN IF NOT EXISTS "resource_version" INTEGER NOT NULL DEFAULT 1 UNIQUE;